base_requirements.txt
pandas==2.2.2
aiohttp==3.9.5
tqdm==4.66.4
loguru==0.7.2
pillow==10.3.0